Issue_377: refactoring
